etl_services/requirements.txt
more-itertools==8.14.0
psycopg2-binary==2.9
pydantic==1.9.1
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.1
orjson==3.7.12